Added MovieWithRating class to return both fields.
My html file only had the attribute of a movie object, and my repository method:"findByMovieRatingAverageRatingGreaterThan(Double rating);"
would return an array of movies with a rating greater than the user input specified . Whilst the movies returned were correct, the movie object
does not have any fields for rating hence I would have to do movie.getMovieRating().getAverageRating() to get the corresponding movie object's
averagerating field. And I couldn't call this after the repository method was called becase all the movie objects are stored in an array list.

I decided a way I could tackle this was to return a new class hybrid of the two objects Movie and MovieRating and make that
the object that is returned in an array list when the repo query method is called.

But when I did that I had a toString problem and didn't know how to access the movie object's and movierating's object's fields in the toString
so I just made MovieWithRating take in two parameters and they would be String (primaryTitle of the movie object), and Double (avgrating of the movie object),
then I could call a toString and format it to use the primary title and avgrating fields after the constructor intializes those fields.